Move URL migration fns to GrapherUrlMigrations.ts
To create a convention, e.g. ExplorerUrlMigrations.ts
