+++ grapher/core/GrapherInterface.ts
@@ -23,10 +23,7 @@ import {
 } from "../../coreTable/CoreTableConstants"
 import { EntityId, EntityName } from "../../coreTable/OwidTableConstants"
 import { ColorSchemeName } from "../color/ColorConstants"
-import { EntityUrlBuilder } from "./EntityUrlBuilder"
 import { QueryParams } from "../../clientUtils/url"
-import { Url } from "../../urls/Url"
-import { UrlMigration, performUrlMigrations } from "../../urls/UrlMigration"
 
 // This configuration represents the entire persistent state of a grapher
 // Ideally, this is also all of the interaction state: when a grapher is saved and loaded again
@@ -173,32 +170,3 @@ export const grapherKeysToSerialize = [
     "comparisonLines",
     "relatedQuestions",
 ]
-
-export const grapherUrlMigrations: UrlMigration[] = [
-    (url) => {
-        const { year, time } = url.queryParams
-        if (!year) return url
-        return url.updateQueryParams({
-            year: undefined,
-            time: time ?? year,
-        })
-    },
-    (url) => {
-        const { country } = url.queryParams
-        if (!country) return url
-        return url.updateQueryParams({
-            country: undefined,
-            selection: EntityUrlBuilder.migrateLegacyCountryParam(country!),
-        })
-    },
-]
-
-export const legacyToCurrentGrapherUrl = (url: Url) =>
-    performUrlMigrations(grapherUrlMigrations, url)
-
-export const legacyToCurrentGrapherQueryParams = (
-    params: LegacyGrapherQueryParams
-): QueryParams => {
-    const url = Url.fromQueryParams(params)
-    return legacyToCurrentGrapherUrl(url).queryParams
-}
